Slider: prefer stof to stoi
When converting the string value, prefer stof to handle both floating
and integer cases. The implicit conversion makes the trick for the
integer case.

+++ include/egt/slider.h
@@ -674,7 +674,7 @@ void SliderType<T>::deserialize(const std::string& name, const std::string& valu
     else if (name == "ending")
         this->ending(std::stoi(value));
     else if (name == "value")
-        this->value(std::stoi(value));
+        this->value(std::stof(value));
     else
         ValueRangeWidget<T>::deserialize(name, value, attrs);
 }
